requirements.txt
tensorflow>=2.10.0
opencv-python>=4.6.0
streamlit>=1.15.0
numpy>=1.23.0
pandas>=1.5.0
matplotlib>=3.6.0
pillow>=9.3.0
scikit-learn>=1.1.0
seaborn>=0.12.0
albumentations>=1.3.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
